rook: Restrict input sampling rate to 16000
It seems that Amazon completely messed up the Audio HAL to the
point where MediaTeks included solution for resampling captured
audio to the requested rate (libblisrc32) does not work at all,
both up-sampling and down-sampling produce incorrect results.

As a workaround, restrict the input sampling rate to match the
hardware, which is 16000.

Thanks to @shortcircuitguru on xda-developers for the hint!

Change-Id: If383653836b0fdc1f6da68942f8d13873b6461d2
Signed-off-by: bengris32 <bengris32@protonmail.ch>